refactor: controller로 view 패키지가 의존되도록 수정
