+++ src/main/kotlin/racing/controller/RacingGame.kt
@@ -1,19 +1,19 @@
-package racing.application
+package racing.controller
 
 import racing.domain.Car
-import racing.domain.CarName
 import racing.domain.Cars
 import racing.domain.RandomStrategy
-import racing.ui.OutputView.printCarNameAndPosition
-import racing.ui.OutputView.printTitle
-import racing.ui.OutputView.printWinnerCars
+import racing.view.InputView
+import racing.view.OutputView.printCarNameAndPosition
+import racing.view.OutputView.printTitle
+import racing.view.OutputView.printWinnerCars
 
-class RacingGame(
-    val tryCount: Int,
-    val carNames: List<CarName>
-) {
+class RacingGame{
 
     fun runRace() {
+        val carNames = InputView.getCarNames()
+        val tryCount = InputView.getTryCount()
+
         val cars = Cars(carNames.map { Car(it) }, RandomStrategy())
         printTitle()
